requirements.txt
PyJWT==2.8.0
Requests==2.31.0
SQLAlchemy==1.4.50
fastapi==0.104.1
gunicorn==21.2.0
passlib==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
python_jose==3.3.0
uvicorn==0.24.0.post1
